Fix pixel_offset rounding-error calculation that could cause minor corruption while scrolling.
git-svn-id: https://xchat.svn.sourceforge.net/svnroot/xchat@452 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1

+++ src/fe-gtk/xtext.c
@@ -4014,7 +4014,7 @@ gtk_xtext_render_page (GtkXText * xtext)
 		return;
 
 #ifdef SMOOTH_SCROLL
-	xtext->pixel_offset = ((float)((float)xtext->adj->value - (float)startline) * xtext->fontsize);
+	xtext->pixel_offset = (xtext->adj->value - startline) * xtext->fontsize;
 #else
 	xtext->pixel_offset = 0;
 #endif
